+++ src/classes/Car.js
@@ -1,0 +1,6 @@
+export default class Car {
+  constructor(name) {
+    this.name = name;
+    this.point = 0;
+  }
+}
