+++ src/index.js
@@ -1,23 +1,25 @@
 import { $ } from './utils/selector.js';
 import { submitCarNames } from './utils/submitCarNames.js';
 import { submitRacingCount } from './utils/submitRacingCount.js';
+import { playRacingGame } from './utils/playRacingGame.js';
 
 export default class RacingGame {
   constructor() {
-    this.carNames = [];
+    this.cars = [];
     this.racingCount = 0;
     this.initEventListeners();
   }
 
   initEventListeners() {
     $('#car-names-submit').addEventListener('click', () => {
-      const carNames = submitCarNames();
-      console.log(carNames);
+      this.cars = submitCarNames();
     });
 
     $('#racing-count-submit').addEventListener('click', () => {
-      const racingCount = submitRacingCount();
-      console.log(racingCount);
+      this.racingCount = submitRacingCount();
+      if (this.cars.length && this.racingCount) {
+        playRacingGame(this.cars, this.racingCount);
+      }
     });
   }
 }
