+++ src/utils/playRacingGame.js
@@ -1,0 +1,3 @@
+export const playRacingGame = (cars, racingCount) => {
+  console.log(cars, racingCount);
+};
